+++ src/ThumbCreator.php
@@ -18,11 +18,8 @@ use Intervention\Image\Constraint;
 use Intervention\Image\Exception\NotReadableException;
 use Intervention\Image\Image;
 use Intervention\Image\ImageManager;
+use LogicException;
 use Symfony\Component\Filesystem\Exception\IOException;
-use Thumber\Exception\NotReadableImageException;
-use Thumber\Exception\UnsupportedImageTypeException;
-use Tools\Exception\NotWritableException;
-use Tools\Exceptionist;
 use Tools\Filesystem;
 
 /**
@@ -75,11 +72,14 @@ class ThumbCreator
      *
      * It sets the file path and extension.
      * @param string $path Path of the image from which to create the thumbnail. It can be a full path or a remote url
-     * @throws \Tools\Exception\NotReadableException
+     * @throws \LogicException
      */
     public function __construct(string $path)
     {
-        $this->path = is_url($path) ? $path : Exceptionist::isReadable($path);
+        if (!is_url($path) && !is_readable($path)) {
+            throw new LogicException('File or directory `' . $path . '` is not readable');
+        }
+        $this->path = $path;
         $this->Filesystem = new Filesystem();
         $this->ImageManager = new ImageManager(['driver' => THUMBER_DRIVER]);
         $this->arguments[] = $path;
@@ -87,7 +87,7 @@ class ThumbCreator
 
     /**
      * Internal method to get default options for the `save()` method
-     * @param array<string, mixed> $options Passed options
+     * @param array $options Passed options
      * @param string $path Path to use
      * @return array{format: string, quality: int, target: false|string} Passed options with default options
      */
@@ -110,20 +110,19 @@ class ThumbCreator
      * Gets an `Image` instance
      * @return \Intervention\Image\Image
      * @throws \ErrorException
-     * @throws \Thumber\Exception\NotReadableImageException
-     * @throws \Thumber\Exception\UnsupportedImageTypeException
      */
     protected function getImageInstance(): Image
     {
         try {
             $ImageInstance = $this->ImageManager->make($this->path);
         } catch (NotReadableException $e) {
+            $path = $this->Filesystem->rtr($this->path) ?: null;
+            $message = $path ? 'Unable to read image from `' . $path . '`' : 'Unable to read image from file';
             if (str_starts_with($e->getMessage(), 'Unsupported image type')) {
                 $type = mime_content_type($this->path) ?: null;
-                throw new UnsupportedImageTypeException($type ? 'Image type `' . $type . '` is not supported by this driver' : 'Image type not supported by this driver');
+                $message = $type ? 'Image type `' . $type . '` is not supported by this driver' : 'Image type not supported by this driver';
             }
-            $path = $this->Filesystem->rtr($this->path) ?: null;
-            throw new NotReadableImageException($path ? 'Unable to read image from `' . $path . '`' : 'Unable to read image from file');
+            throw new LogicException($message);
         }
 
         return $ImageInstance;
@@ -135,22 +134,31 @@ class ThumbCreator
      * You can define optional coordinates to move the top-left corner of the cutout to a certain position.
      * @param int $width Required width
      * @param int $height Required height. If not specified, the width will be used
-     * @param array<string, int> $options Options for the thumbnail
+     * @param array $options Options for the thumbnail
      * @return self
+     * @throws \LogicException
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#crop
      */
     public function crop(int $width, int $height = 0, array $options = []): ThumbCreator
     {
         $height = $height ?: $width;
+        if (!is_positive($width)) {
+            throw new LogicException(sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
+        }
         $options += ['x' => 0, 'y' => 0];
-        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
+        if (!is_int($options['x'])) {
+            throw new LogicException('The `x` option must be an integer');
+        }
+        if (!is_int($options['y'])) {
+            throw new LogicException('The `y` option must be an integer');
+        }
 
         $this->arguments[] = [__FUNCTION__, $width, $height, $options];
         $this->callbacks[] = fn(Image $ImageInstance): Image => $ImageInstance->crop(
             $width,
             $height,
-            Exceptionist::isInt($options['x'], 'The `x` option must be an integer'),
-            Exceptionist::isInt($options['y'], 'The `y` option must be an integer')
+            $options['x'],
+            $options['y'],
         );
 
         return $this;
@@ -161,14 +169,16 @@ class ThumbCreator
      *  aspect ratio on the current image automatically, cut it out and resize it to the given dimension
      * @param int $width Required width. If not specified, the height will be used
      * @param int $height Required height. If not specified, the width will be used
-     * @param array<string, mixed> $options Options for the thumbnail
+     * @param array $options Options for the thumbnail
      * @return self
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#fit
      */
     public function fit(int $width = 0, int $height = 0, array $options = []): ThumbCreator
     {
         $width = $width ?: $height;
-        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
+        if (!is_positive($width)) {
+            throw new LogicException(sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
+        }
         $height = $height ?: $width;
         $options += ['position' => 'center', 'upsize' => true];
 
@@ -187,15 +197,18 @@ class ThumbCreator
      * Resizes the image
      * @param int $width Required width
      * @param int $height Required height. If not specified, the width will be used
-     * @param array<string, bool> $options Options for the thumbnail
+     * @param array $options Options for the thumbnail
      * @return self
+     * @throws \LogicException
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#resize
      */
     public function resize(int $width, int $height = 0, array $options = []): ThumbCreator
     {
         $height = $height ?: $width;
+        if (!is_positive($width)) {
+            throw new LogicException(sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
+        }
         $options += ['aspectRatio' => true, 'upsize' => true];
-        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
 
         $this->arguments[] = [__FUNCTION__, $width, $height, $options];
         $this->callbacks[] = fn(Image $ImageInstance): Image => $ImageInstance
@@ -217,15 +230,18 @@ class ThumbCreator
      *  width or height to the actual image dimensions. You can also pass a background color for the emerging area of the image
      * @param int $width Required width
      * @param int $height Required height. If not specified, the width will be used
-     * @param array<string, mixed> $options Options for the thumbnail
+     * @param array $options Options for the thumbnail
      * @return self
+     * @throws \LogicException
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#resizecanvas
      */
     public function resizeCanvas(int $width, int $height = 0, array $options = []): ThumbCreator
     {
         $height = $height ?: $width;
+        if (!is_positive($width)) {
+            throw new LogicException(sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
+        }
         $options += ['anchor' => 'center', 'relative' => false, 'bgcolor' => '#ffffff'];
-        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__));
 
         $this->arguments[] = [__FUNCTION__, $width, $height, $options];
         $this->callbacks[] = fn(Image $ImageInstance): Image => $ImageInstance
@@ -236,16 +252,16 @@ class ThumbCreator
 
     /**
      * Saves the thumbnail and returns its path
-     * @param array<string, mixed> $options Options for saving
+     * @param array $options Options for saving
      * @return string Thumbnail path
-     * @throws \Tools\Exception\NotWritableException
-     * @throws \Thumber\Exception\NotReadableImageException
      * @throws \ErrorException
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#save-the-thumbnail
      */
     public function save(array $options = []): string
     {
-        Exceptionist::isTrue($this->callbacks, 'No valid method called before the `save()` method');
+        if (!$this->callbacks) {
+            throw new LogicException('No valid method called before the `save()` method');
+        }
 
         $options = $this->getDefaultSaveOptions($options);
         $target = $options['target'];
@@ -271,7 +287,7 @@ class ThumbCreator
             try {
                 $this->Filesystem->dumpFile($target, (string)$content);
             } catch (IOException $e) {
-                throw new NotWritableException(sprintf('Unable to create file `%s`', $target ? $this->Filesystem->rtr($target) : ''));
+                throw new LogicException(sprintf('Unable to create file `%s`', $target ? $this->Filesystem->rtr($target) : ''));
             }
         }