+++ tests/TestCase/ThumbCreatorTest.php
@@ -17,16 +17,16 @@ namespace Thumber\Test;
 
 use Intervention\Image\Exception\NotReadableException as InterventionNotReadableException;
 use Intervention\Image\ImageManager;
-use Thumber\Exception\NotReadableImageException;
-use Thumber\Exception\UnsupportedImageTypeException;
 use Thumber\TestSuite\TestCase;
-use Tools\Exception\NotReadableException;
+use Tools\TestSuite\ReflectionTrait;
 
 /**
  * ThumbCreatorTest class
  */
 class ThumbCreatorTest extends TestCase
 {
+    use ReflectionTrait;
+
     /**
      * Test for `__construct()` method, passing a no existing file
      * @test
@@ -34,7 +34,7 @@ class ThumbCreatorTest extends TestCase
      */
     public function testConstructNoExistingFile(): void
     {
-        $this->expectException(NotReadableException::class);
+        $this->expectExceptionMessageMatches('/^File or directory `[\w\-\/\.\\\:]+` is not readable$/');
         $this->getThumbCreatorInstance('noExistingFile.gif');
     }
 
@@ -45,14 +45,7 @@ class ThumbCreatorTest extends TestCase
      */
     public function testGetImageInstanceNotReadableImageException(): void
     {
-        $expectedException = NotReadableImageException::class;
-        $expectedMessage = 'Unable to read image from `tests/bootstrap.php`';
-        if (THUMBER_DRIVER != 'imagick') {
-            $expectedException = UnsupportedImageTypeException::class;
-            $expectedMessage = 'Image type `text/x-php` is not supported by this driver';
-        }
-        $this->expectException($expectedException);
-        $this->expectExceptionMessage($expectedMessage);
+        $this->expectExceptionMessage(THUMBER_DRIVER != 'imagick' ? 'Image type `text/x-php` is not supported by this driver' : 'Unable to read image from `tests/bootstrap.php`');
         $this->getThumbCreatorInstanceWithSave(TESTS . 'bootstrap.php');
     }
 
@@ -63,7 +56,6 @@ class ThumbCreatorTest extends TestCase
      */
     public function testGetImageInstanceUnsupportedImageType(): void
     {
-        $this->expectException(UnsupportedImageTypeException::class);
         $this->expectExceptionMessage('Image type `image/jpeg` is not supported by this driver');
         $exception = new InterventionNotReadableException('Unsupported image type. GD driver is only able to decode JPG, PNG, GIF or WebP files.');
         $thumbCreator = $this->getThumbCreatorInstance();